feat: new routing method to dynamic url
